Added ControlTarget composition for Annotation and Controllable.
