+++ ts/Extensions/Annotations/AnnotationOptions.d.ts
@@ -19,10 +19,10 @@ import type Annotation from './Annotation';
 import type AST from '../../Core/Renderer/HTML/AST';
 import type {
     ControllableLabelOptions,
-    ControllableOptions,
     ControllableShapeOptions
 } from './Controllables/ControllableOptions';
-import type { ControlPointOptionsObject } from './ControlPointOptions';
+import type ControlPointOptions from './ControlPointOptions';
+import type ControlTargetOptions from './ControlTargetOptions';
 import type EventCallback from '../../Core/EventCallback';
 import type MockPointOptions from './MockPointOptions';
 import type NavigationOptions from '../Exporting/NavigationOptions';
@@ -45,9 +45,9 @@ export interface AnnotationEventsOptions {
     touchend?: EventCallback<Annotation>;
 }
 
-export interface AnnotationOptions extends ControllableOptions { // @todo AnnotationOptions.d.ts
+export interface AnnotationOptions extends ControlTargetOptions {
     animation: Partial<AnimationOptions>;
-    controlPointOptions: ControlPointOptionsObject;
+    controlPointOptions: ControlPointOptions;
     crop: boolean;
     draggable: AnnotationDraggableValue;
     events: AnnotationEventsOptions;
@@ -56,8 +56,6 @@ export interface AnnotationOptions extends ControllableOptions { // @todo Annota
     labelOptions?: ControllableLabelOptions;
     labels?: Array<ControllableLabelOptions>;
     langKey?: string;
-    point?: (string|MockPointOptions);
-    points?: Array<(string|MockPointOptions)>;
     shapeOptions: ControllableShapeOptions;
     shapes?: Array<ControllableShapeOptions>;
     type?: string;
