+++ ts/Extensions/Annotations/Controllables/Controllable.ts
@@ -15,22 +15,14 @@
 import type Annotation from '../Annotation';
 import type AnnotationChart from '../AnnotationChart';
 import type { AnnotationPointType } from '../AnnotationSeries';
-import type BBoxObject from '../../../Core/Renderer/BBoxObject';
 import type ControllableLike from './ControllableLike';
 import type ControllableOptions from './ControllableOptions';
-import type MockPointOptions from '../MockPointOptions';
 import type SVGAttributes from '../../../Core/Renderer/SVG/SVGAttributes';
 import type SVGElement from '../../../Core/Renderer/SVG/SVGElement';
 
-import ControlPoint from '../ControlPoint.js';
-import MockPoint from '../MockPoint.js';
+import ControlTarget from '../ControlTarget.js';
 import U from '../../../Core/Utilities.js';
-const {
-    isObject,
-    isString,
-    merge,
-    splat
-} = U;
+const { merge } = U;
 
 /* *
  *
@@ -39,11 +31,6 @@ const {
  * */
 
 export type AttrsMapObject = Record<keyof ControllableOptions, keyof SVGAttributes>;
-
-export interface ControllableAnchorObject {
-    absolutePosition: BBoxObject;
-    relativePosition: BBoxObject;
-}
 
 /* *
  *
@@ -56,7 +43,7 @@ export interface ControllableAnchorObject {
  * and points transformations.
  * @private
  */
-abstract class Controllable {
+abstract class Controllable implements ControlTarget {
 
     /* *
      *
@@ -90,12 +77,10 @@ abstract class Controllable {
     public annotation: Annotation;
     public chart: AnnotationChart;
     public collection: ('labels'|'shapes');
-    public controlPoints: Array<ControlPoint>;
     public graphic: SVGElement = void 0 as any;
     public index: number;
     public itemType: ('label'|'shape');
     public options: ControllableOptions;
-    public points: Array<AnnotationPointType>;
     public tracker?: SVGElement;
 
     /* *
@@ -103,73 +88,6 @@ abstract class Controllable {
      *  Functions
      *
      * */
-
-    /**
-     * Add control points to a controllable.
-     * @private
-     */
-    public addControlPoints(): void {
-        const controlPoints = this.controlPoints,
-            controlPointsOptions = this.options.controlPoints || [];
-
-        controlPointsOptions.forEach((controlPointOptions, i): void => {
-            const options = merge(
-                this.options.controlPointOptions,
-                controlPointOptions
-            );
-
-            if (!options.index) {
-                options.index = i;
-            }
-
-            controlPointsOptions[i] = options;
-
-            controlPoints.push(new ControlPoint(
-                this.chart,
-                this,
-                options
-            ));
-        });
-    }
-
-    /**
-     * Returns object which denotes anchor position - relative and absolute.
-     * @private
-     * @param {Highcharts.AnnotationPointType} point
-     *        A point like object.
-     * @return {Highcharts.AnnotationAnchorObject}
-     *         A controllable anchor
-     */
-    public anchor(
-        point: AnnotationPointType
-    ): ControllableAnchorObject {
-        const plotBox = point.series.getPlotBox(),
-            chart = point.series.chart,
-            box = point.mock ?
-                point.toAnchor() :
-                chart.tooltip &&
-                chart.tooltip.getAnchor.call({
-                    chart: point.series.chart
-                }, point) ||
-                [0, 0, 0, 0],
-
-            anchor = {
-                x: box[0] + (this.options.x || 0),
-                y: box[1] + (this.options.y || 0),
-                height: box[2] || 0,
-                width: box[3] || 0
-            };
-
-        return {
-            relativePosition: anchor,
-            absolutePosition: merge(anchor, {
-                x: anchor.x + (
-                    point.mock ? plotBox.translateX : chart.plotLeft
-                ),
-                y: anchor.y + (point.mock ? plotBox.translateY : chart.plotTop)
-            })
-        };
-    }
 
     /**
      * Redirect attr usage on the controllable graphic element.
@@ -245,21 +163,6 @@ abstract class Controllable {
     }
 
     /**
-     * Get the controllable's points options.
-     * @private
-     * @return {Array<Highcharts.PointOptionsObject>}
-     * An array of points' options.
-     */
-    public getPointsOptions(): Array<MockPointOptions> {
-        const options = this.options;
-
-        return (
-            options.points ||
-            (options.point && splat(options.point))
-        ) as any;
-    }
-
-    /**
      * Init the controllable
      * @private
      */
@@ -277,83 +180,6 @@ abstract class Controllable {
 
         this.linkPoints();
         this.addControlPoints();
-    }
-
-    /**
-     * Find point-like objects based on points options.
-     * @private
-     * @return {Array<Annotation.PointLike>}
-     *         An array of point-like objects.
-     */
-    public linkPoints(): (Array<AnnotationPointType>|undefined) {
-        const pointsOptions = this.getPointsOptions(),
-            points = this.points,
-            len = (pointsOptions && pointsOptions.length) || 0;
-
-        let i: number,
-            point: (AnnotationPointType|null);
-
-        for (i = 0; i < len; i++) {
-            point = this.point(pointsOptions[i], points[i]);
-
-            if (!point) {
-                points.length = 0;
-
-                return;
-            }
-
-            if (point.mock) {
-                point.refresh();
-            }
-
-            points[i] = point;
-        }
-
-        return points;
-    }
-
-    /**
-     * Map point's options to a point-like object.
-     * @private
-     * @param {string|Function|Highcharts.AnnotationMockPointOptionsObject|Highcharts.AnnotationPointType} pointOptions
-     *        Point's options.
-     * @param {Highcharts.AnnotationPointType} point
-     *        A point-like instance.
-     * @return {Highcharts.AnnotationPointType|null}
-     *         If the point is found/set returns this point, otherwise null
-     */
-    public point(
-        pointOptions: (string|Function|MockPoint|MockPointOptions),
-        point: (AnnotationPointType|null)
-    ): (AnnotationPointType|null) {
-        if (pointOptions && (pointOptions as MockPointOptions).series) {
-            return pointOptions as any;
-        }
-
-        if (!point || point.series === null) {
-            if (isObject(pointOptions)) {
-                point = new MockPoint(
-                    this.chart,
-                    this,
-                    pointOptions as MockPointOptions
-                );
-            } else if (isString(pointOptions)) {
-                point = (this.chart.get(pointOptions) as any) || null;
-            } else if (typeof pointOptions === 'function') {
-                const pointConfig: (MockPoint|MockPointOptions) =
-                    pointOptions.call(point, this);
-
-                point = pointConfig.series ?
-                    pointConfig :
-                    new MockPoint(
-                        this.chart,
-                        this,
-                        pointOptions
-                    );
-            }
-        }
-
-        return point;
     }
 
     /**
@@ -441,107 +267,6 @@ abstract class Controllable {
     }
 
     /**
-     * Transform a controllable with a specific transformation.
-     * @private
-     * @param {string} transformation
-     *        A transformation name
-     * @param {number|null} cx
-     *        Origin x transformation
-     * @param {number|null} cy
-     *        Origin y transformation
-     * @param {number} p1
-     *        Param for the transformation
-     * @param {number} [p2]
-     *        Param for the transformation
-     */
-    public transform(
-        transformation: string,
-        cx: (number|null),
-        cy: (number|null),
-        p1: number,
-        p2?: number
-    ): void {
-        if (this.chart.inverted) {
-            const temp = cx;
-
-            cx = cy;
-            cy = temp;
-        }
-
-        this.points.forEach((_point, i): void => (
-            this.transformPoint(transformation, cx, cy, p1, p2, i)
-        ), this);
-    }
-
-    /**
-     * Transform a point with a specific transformation
-     * If a transformed point is a real point it is replaced with
-     * the mock point.
-     * @private
-     * @param {string} transformation
-     *        A transformation name
-     * @param {number|null} cx
-     *        Origin x transformation
-     * @param {number|null} cy
-     *        Origin y transformation
-     * @param {number} p1
-     *        Param for the transformation
-     * @param {number|undefined} p2
-     *        Param for the transformation
-     * @param {number} i
-     *        Index of the point
-     */
-    public transformPoint(
-        transformation: string,
-        cx: (number|null),
-        cy: (number|null),
-        p1: number,
-        p2: (number|undefined),
-        i: number
-    ): void {
-        let point = this.points[i];
-
-        if (!point.mock) {
-            point = this.points[i] = MockPoint.fromPoint(point);
-        }
-
-        (point as any)[transformation](cx, cy, p1, p2);
-    }
-
-    /**
-     * Translate a controllable.
-     * @private
-     * @param {number} dx
-     *        Translation for x coordinate
-     * @param {number} dy
-     *        Translation for y coordinate
-     **/
-    public translate(
-        dx: number,
-        dy: number
-    ): void {
-        this.transform('translate', null, null, dx, dy);
-    }
-
-    /**
-     * Translate a specific point within a controllable.
-     * @private
-     * @param {number} dx
-     *        Translation for x coordinate
-     * @param {number} dy
-     *        Translation for y coordinate
-     * @param {number} i
-     *        Index of the point
-     **/
-    public translatePoint(
-        dx: number,
-        dy: number,
-        i: number
-    ): void {
-        this.transformPoint('translate', null, null, dx, dy, i);
-    }
-
-    /**
      * Translate shape within controllable item.
      * Replaces `controllable.translate` method.
      * @private
@@ -613,9 +338,11 @@ abstract class Controllable {
  *
  * */
 
-interface Controllable extends ControllableLike {
+interface Controllable extends ControllableLike, ControlTarget {
     // placeholder for additional class members
 }
+
+ControlTarget.compose(Controllable);
 
 /* *
  *
