+++ ts/Extensions/Annotations/Controllables/ControllableOptions.d.ts
@@ -20,14 +20,13 @@ import type {
 } from '../../../Core/Renderer/AlignObject';
 import type { AnnotationPoint } from '../AnnotationSeries';
 import type ColorType from '../../../Core/Color/ColorType';
-import type { ControlPointOptionsObject } from '../ControlPointOptions';
 import type CSSObject from '../../../Core/Renderer/CSSObject';
+import type ControlTargetOptions from '../ControlTargetOptions';
 import type DashStyleValue from '../../../Core/Renderer/DashStyleValue';
 import type {
     DataLabelOverflowValue
 } from '../../../Core/Series/DataLabelOptions';
 import type FormatUtilities from '../../../Core/FormatUtilities';
-import type MockPointOptions from '../MockPointOptions';
 import type {
     ShadowOptionsObject
 } from '../../../Core/Renderer/ShadowOptionsObject';
@@ -68,21 +67,15 @@ export interface ControllableLabelOptions extends ControllableOptions {
     yAxis?: number|string;
 }
 
-export interface ControllableOptions {
+export interface ControllableOptions extends ControlTargetOptions {
     className?: string;
-    controlPointOptions?: ControlPointOptionsObject;
-    controlPoints?: Array<ControlPointOptionsObject>;
     id?: (number|string);
     markerEnd?: string;
     markerStart?: string;
-    point?: (string|MockPointOptions);
-    points?: Array<(string|MockPointOptions)>;
     r?: number;
     rx?: number;
     ry?: number;
     type?: string;
-    x?: number;
-    y?: number;
 }
 
 export interface ControllableShapeOptions extends ControllableOptions {
