+++ ts/Extensions/Annotations/Types/BasicAnnotation.ts
@@ -22,7 +22,7 @@ import type {
 } from '../Controllables/ControllableOptions';
 import type ControllableRect from '../Controllables/ControllableRect';
 import type ControlPoint from '../ControlPoint';
-import type { ControlPointOptionsObject } from '../ControlPointOptions';
+import type ControlPointOptions from '../ControlPointOptions';
 import type MockPointOptions from '../MockPointOptions';
 import type PointerEvent from '../../../Core/PointerEvent';
 import type PositionObject from '../../../Core/Renderer/PositionObject';
@@ -71,7 +71,7 @@ class BasicAnnotation extends Annotation {
                 drag: function (
                     this: Annotation,
                     e: AnnotationEventObject,
-                    target: Annotation
+                    target: Controllable
                 ): void {
                     const xy = this.mouseMoveToTranslation(e);
 
@@ -383,10 +383,10 @@ BasicAnnotation.prototype.defaultOptions = merge(
 
 namespace BasicAnnotation {
     export interface ControlPoints {
-        label: DeepPartial<ControlPointOptionsObject>[];
-        rectangle: DeepPartial<ControlPointOptionsObject>[];
-        ellipse: DeepPartial<ControlPointOptionsObject>[];
-        circle: DeepPartial<ControlPointOptionsObject>[];
+        label: DeepPartial<ControlPointOptions>[];
+        rectangle: DeepPartial<ControlPointOptions>[];
+        ellipse: DeepPartial<ControlPointOptions>[];
+        circle: DeepPartial<ControlPointOptions>[];
     }
 }
